+++ examples/blink-led.js
@@ -1,13 +1,17 @@
-var onoff = require('../onoff'),
-    ledGpio = 17,
-    nextLedState = 1;
+var Gpio = require('../onoff').Gpio, // Constructor function for Gpio objects.
+    ledGpio = new Gpio(17, 'out'),   // Export GPIO #17 as an output.
+    iv;
 
-onoff.configure(ledGpio, 'out', function (err) {
-    if (err) throw err;
+// Toggle the state of the LED on GPIO #17 every 200ms.
+// Here synchronous methods are used. Asynchronous methods are also available.
+iv = setInterval(function() {
+    ledGpio.writeSync(ledGpio.readSync() === 0 ? 1 : 0); // 1 = on, 0 = off.
+}, 200);
 
-    setInterval(function () {
-        onoff.value(ledGpio, nextLedState);
-        nextLedState = nextLedState === 1 ? 0 : 1;
-    }, 200);
-});
+// Stop blinking the LED and turn it off after 5 seconds.
+setTimeout(function() {
+    clearInterval(iv);    // Stop blinking
+    ledGpio.writeSync(0); // Turn LED off.
+    ledGpio.unexport();   // Unexport GPIO and free resources
+}, 5000);