+++ examples/no-sudo/export-gpios.js
@@ -1,0 +1,3 @@
+var Gpio = require('../../onoff').Gpio,
+    ledGpio = new Gpio(/* 38 */ 17, 'out');
+