+++ examples/no-sudo/unexport-gpios.js
@@ -1,0 +1,5 @@
+var Gpio = require('../../onoff').Gpio,
+    ledGpio = new Gpio(/* 38 */ 17, 'out');
+
+ledGpio.unexport();
+