+++ examples/wait-for-interrupt.js
@@ -1,13 +1,16 @@
-var onoff = require('../onoff'),
-    buttonGpio = 18;
+var Gpio = require('../onoff').Gpio,  // Constructor function for Gpio objects.
+    buttonGpio = new Gpio(18, 'in', 'both'); // Export GPIO #18 as an interrupt
+                                             // generating input.
 
-onoff.configure(buttonGpio, 'in', 'both', function (err) {
+console.log('Please press the button on GPIO #18...');
+
+// The callback passed to watch will be called when the button on GPIO #18 is
+// pressed. 
+buttonGpio.watch(function (err, value) {
     if (err) throw err;
-    console.log('Please press the button on GPIO 18...');
-    onoff.watch(buttonGpio, function (err, value) {
-        if (err) throw err;
-        console.log('Button pressed!, its value was ' + value);
-        onoff.unexp(buttonGpio);
-    });
+
+    console.log('Button pressed!, its value was ' + value);
+
+    buttonGpio.unexport(); // Unexport GPIO and free resources
 });