+++ History.md
@@ -1,4 +1,14 @@
+0.1.0 - Nov ?? 2012
+===================
+
+  * Added Gpio objects
+  * Removed functions, use Gpio objects instead
+  * Performance improvements
+  * Synchronous or asynchronous access to a GPIOs value
+  * Allow applications to handle superuser issues
+
 0.0.1 - Oct 28 2012
 ===================
 
   * Initial release
+